requirements.txt
# FastAPI and related dependencies
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6

# Database (PostgreSQL + SQLAlchemy)
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
asyncpg==0.27.0

# Machine Learning (TensorFlow)
tensorflow==2.12.0

# Image Processing (OpenCV)
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72  # For headless environments

# OCR (Tesseract)
pytesseract==0.3.10
pillow==9.5.0

# Geolocation (PostGIS)
geoalchemy2==0.14.0
shapely==2.0.1

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1





























































# FastAPI and related dependencies
fastapi==0.115.8
uvicorn==0.32.1
python-multipart==0.0.20
fastapi-mail==1.4.2
email-validator==2.2.0

# Database (PostgreSQL + SQLAlchemy)
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Machine Learning (TensorFlow)
# tensorflow==2.18.0
tensorflow-cpu==2.15.0

tensorflow-io-gcs-filesystem==0.31.0
# tensorboard==2.18.0
tensorboard>=2.15,<2.16
tensorboard-data-server==0.7.2
tf_keras==2.18.0
# keras==3.6.0
keras>=2.15.0,<2.16 
# ml-dtypes==0.4.1
# wrapt==1.16.0
wrapt>=1.11.0,<1.15 
gast==0.6.0
opt_einsum==3.4.0
flatbuffers==24.3.25
# protobuf==5.28.3
protobuf>=3.20.3,<5.0.0

absl-py==2.1.0
termcolor==2.5.0
google-pasta==0.2.0
libclang==18.1.1

# Image Processing (OpenCV)
opencv-python==4.11.0.86
opencv-python-headless==4.7.0.72  # For headless environments

# OCR (Tesseract)
pytesseract==0.3.10
pillow==11.0.0

# Geolocation (PostGIS)
geoalchemy2==0.14.0
shapely==2.0.1
geopy==2.4.1
geographiclib==2.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.28.1

# Session middleware and related
itsdangerous==2.2.0

# Password hashing and JWT
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# Pydantic
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0

# Web server framework
starlette==0.41.3
asgiref==3.8.1

# Email
aiosmtplib==3.0.2

# Utility libraries
click==8.1.7
python-dateutil==2.9.0.post0
typing_extensions==4.12.2
packaging==24.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3
anyio==4.6.2.post1
sniffio==1.3.1
email_validator==2.2.0

# Security
bcrypt==4.3.0
cryptography==43.0.3
cffi==1.17.1
pycparser==2.22

# File system and serialization
aiofiles==24.1.0
orjson==3.10.15
ujson==5.10.0

# PDF and document processing
PyPDF2==3.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
python-docx==1.1.2

# Optional web tools
Jinja2==3.1.5
MarkupSafe==3.0.2

# CLI tools
# typer==0.13.0
typer==0.9.0
rich==13.9.4

# Spacy and NLP
# spacy==3.8.2
spacy==3.5.4

# blis==1.0.1
blis==0.7.9
catalogue==2.0.10
confection==0.1.5
cymem==2.0.8
langcodes==3.4.1
murmurhash==1.0.10
preshed==3.0.9
srsly==2.4.8
# thinc==8.3.2
thinc==8.1.8

wasabi==1.1.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
tokenizers==0.20.3

# Transformers (HuggingFace)
transformers==4.46.3
huggingface-hub==0.26.2
safetensors==0.4.5

# Torch and vision
torch==2.6.0
torchvision==0.21.0

# Visualization
matplotlib==3.10.1
seaborn==0.13.2

# Pandas and NumPy
pandas==2.2.3
# numpy==2.0.2

numpy==1.24.4

# Others (keep if needed)
scipy==1.14.1
tqdm==4.67.0
joblib==1.4.2
nltk==3.9.1
python-jose==3.4.0
Authlib==1.3.0
ultralytics

# For development/CLI tools
alembic==1.14.1
GitPython==3.1.44
shellingham==1.5.4
watchfiles==1.0.4






















# # FastAPI and related dependencies
# fastapi==0.115.8
# uvicorn==0.32.1
# python-multipart==0.0.20
# fastapi-mail==1.4.2
# email-validator==2.2.0

# # Database (PostgreSQL + SQLAlchemy)
# sqlalchemy==2.0.38
# psycopg2-binary==2.9.10
# asyncpg==0.30.0

# # Machine Learning (TensorFlow)
# tensorflow==2.18.0

# tensorflow-io-gcs-filesystem==0.31.0
# tensorboard==2.18.0
# tensorboard-data-server==0.7.2
# tf_keras==2.18.0
# keras==3.6.0
# ml-dtypes==0.4.1
# wrapt==1.16.0
# gast==0.6.0
# opt_einsum==3.4.0
# flatbuffers==24.3.25
# protobuf==5.28.3

# absl-py==2.1.0
# termcolor==2.5.0
# google-pasta==0.2.0
# libclang==18.1.1

# # Image Processing (OpenCV)
# opencv-python==4.11.0.86
# opencv-python-headless==4.7.0.72  # For headless environments

# # OCR (Tesseract)
# pytesseract==0.3.10
# pillow==11.0.0

# # Geolocation (PostGIS)
# geoalchemy2==0.14.0
# shapely==2.0.1
# geopy==2.4.1
# geographiclib==2.0

# # Environment variables
# python-dotenv==1.0.1

# # Testing
# pytest==7.3.1
# pytest-asyncio==0.21.0
# httpx==0.28.1

# # Session middleware and related
# itsdangerous==2.2.0

# # Password hashing and JWT
# passlib[bcrypt]==1.7.4
# PyJWT==2.10.1

# # Pydantic
# pydantic==2.10.6
# pydantic-core==2.27.2
# pydantic-extra-types==2.10.2
# pydantic-settings==2.8.0

# # Web server framework
# starlette==0.41.3
# asgiref==3.8.1

# # Email
# aiosmtplib==3.0.2

# # Utility libraries
# click==8.1.7
# python-dateutil==2.9.0.post0
# typing_extensions==4.12.2
# packaging==24.2
# certifi==2024.8.30
# charset-normalizer==3.4.0
# idna==3.10
# requests==2.32.3
# urllib3==2.2.3
# anyio==4.6.2.post1
# sniffio==1.3.1
# email_validator==2.2.0

# # Security
# bcrypt==4.3.0
# cryptography==43.0.3
# cffi==1.17.1
# pycparser==2.22

# # File system and serialization
# aiofiles==24.1.0
# orjson==3.10.15
# ujson==5.10.0

# # PDF and document processing
# PyPDF2==3.0.1
# pdfminer.six==20231228
# pdfplumber==0.11.4
# python-docx==1.1.2

# # Optional web tools
# Jinja2==3.1.5
# MarkupSafe==3.0.2

# # CLI tools
# typer==0.13.0
# rich==13.9.4

# # Spacy and NLP
# spacy==3.8.2
# blis==1.0.1
# catalogue==2.0.10
# confection==0.1.5
# cymem==2.0.8
# langcodes==3.4.1
# murmurhash==1.0.10
# preshed==3.0.9
# srsly==2.4.8
# thinc==8.3.2
# wasabi==1.1.3
# spacy-legacy==3.0.12
# spacy-loggers==1.0.5
# tokenizers==0.20.3

# # Transformers (HuggingFace)
# transformers==4.46.3
# huggingface-hub==0.26.2
# safetensors==0.4.5

# # Torch and vision
# torch==2.6.0
# torchvision==0.21.0

# # Visualization
# matplotlib==3.10.1
# seaborn==0.13.2

# # Pandas and NumPy
# pandas==2.2.3
# # numpy==2.0.2


# # Others (keep if needed)
# scipy==1.14.1
# tqdm==4.67.0
# joblib==1.4.2
# nltk==3.9.1
# python-jose==3.4.0
# Authlib==1.3.0
# ultralytics

# # For development/CLI tools
# alembic==1.14.1
# GitPython==3.1.44
# shellingham==1.5.4
# watchfiles==1.0.4